+++ mobile/lib/premium_fortune/presentation/profile/pf_profile_page.dart
@@ -1,3 +1,4 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
@@ -55,7 +56,7 @@ class _PfProfilePageState extends ConsumerState<PfProfilePage> {
                   radius: 40,
                   backgroundColor: PfTheme.purple,
                   backgroundImage:
-                      user.photoUrl != null ? NetworkImage(user.photoUrl!) : null,
+                      user.photoUrl != null ? CachedNetworkImageProvider(user.photoUrl!) : null,
                   child: user.photoUrl == null
                       ? Text(
                           user.displayName[0],